getLocation/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-firestore
flask